fix(crypto): well-define negative-left-shift UB in TweetNaCl-derived car25519/modL
The C99 Ed25519/X25519 field arithmetic (TweetNaCl-derived, signed i64[16]
limbs) left-shifts a negative value in three places — C-standard undefined
behavior, flagged by the new ci_local.sh --sanitize (UBSan) gate on its first
run:

  src/crypto/ed25519/ed25519.c:57  car25519  o[i] -= c << 16   (c can be < 0)
  src/crypto/ed25519/ed25519.c:186 modL      x[j] -= carry << 8 (carry can be < 0)
  src/crypto/x25519/x25519.c:22    car25519  o[i] -= c << 16   (c can be < 0)

TweetNaCl relies on every real compiler implementing `<<` on a negative value
as a 2's-complement shift (which is why the output was already correct and
byte-equal to libsodium), but it is formally UB. Fix: shift as unsigned then
reinterpret — (i64)((uint64_t)c << 16) — bit-identical to the intended
2's-complement result on any conforming platform (C++20 mandates 2's-complement),
so the field output stays byte-equal.

Byte-equality re-proven on GCC 13 post-fix: test-ed25519-c99, test-ed25519-vectors,
test-x25519-c99, and test-c99-vectors (CAVP + libsodium-derived corpus) all pass
under UBSan; the full 31-subcommand consensus+crypto surface is now SAN_FAIL=0.
Scope: these c99 modules serve the wallet + self-tests, not the consensus/daemon
path (still OpenSSL) — but the fix precedes the 1b daemon-crypto migration that
would put them on the chain path. See docs/proofs/DECISION-LOG.md 2026-07-03.

+++ src/crypto/ed25519/ed25519.c
@@ -54,7 +54,11 @@ static void car25519(gf o) {
         o[i] += (1LL << 16);
         c = o[i] >> 16;
         o[(i + 1) * (i < 15)] += c - 1 + 37 * (c - 1) * (i == 15);
-        o[i] -= c << 16;
+        /* c can be negative (signed limb); `c << 16` on a negative value is
+         * UB per C (caught by UBSan). Shift as unsigned then reinterpret:
+         * bit-identical to the intended 2's-complement result on every real
+         * platform, so the field output stays byte-equal to libsodium. */
+        o[i] -= (i64)((uint64_t)c << 16);
     }
 }
 
@@ -183,7 +187,9 @@ static void modL(u8 *r, i64 x[64]) {
         for (j = i - 32; j < i - 12; ++j) {
             x[j] += carry - 16 * x[i] * L[j - (i - 32)];
             carry = (x[j] + 128) >> 8;
-            x[j] -= carry << 8;
+            /* carry may be negative; shift as unsigned (see car25519) —
+             * byte-invariant, but well-defined for UBSan. */
+            x[j] -= (i64)((uint64_t)carry << 8);
         }
         x[j] += carry;
         x[i] = 0;
